backend/requirements.txt
flask
faiss-cpu
numpy
requests
sqlalchemy 
psycopg2-binary 
python-dotenv
pypdf